requirements.txt
requests
clickhouse-connect>=0.6.0
numpy>=1.21.0,<2.0.0
pytz
schedule
python-dotenv
pandas